oakestra_fl/root_layer/root_fl_manager/image_builder_management/github/fl_client_env_builder_image/requirements.txt
requests ~=2.28.1
gitpython ~= 3.1.42
paho-mqtt==2.0
